% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_blm_reg.R
\name{blm.reg.fit}
\alias{blm.reg.fit}
\title{Fit a Bayesian linear model via Gibbs sampling}
\usage{
blm.reg.fit(y, X, Z, prior = list(), control = list())
}
\arguments{
\item{y}{response vector}

\item{X}{fixed effect design matrix, including all the variables that must not be penalized}

\item{Z}{random effect design matrix, including all the variable subject to regularization}

\item{prior}{list of prior parameters. More information are provided in 'Details'.}

\item{control}{list of control parameter. More information are provided in 'Details'.}
}
\value{
Returns an object of class "blm", which is a list containing the following elements:
\describe{
  \item{\code{y}}{response vector}
  \item{\code{X}}{fixed effect design matrix}
  \item{\code{Z}}{random effect design matrix}
  \item{\code{prior}}{list of prior parameters}
  \item{\code{control}}{list of control parameters}
  \item{\code{burn}}{list containig the sampled chains of all the parameters during the burn-in iterations}
  \item{\code{trace}}{list containing the sampled chains of all the parameters after the burn-in period}
  \item{\code{exe.time}}{execution time in seconds}
}
}
\description{
A short description...
}
\details{
Additional details...
}
