% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline_basis.R
\name{convert.to.ospline}
\alias{convert.to.ospline}
\title{Create othogonalized design and penalty matrices for O'Sullivan spline regression}
\usage{
convert.to.ospline(x, B = NULL, P = NULL, check = TRUE)
}
\description{
Create othogonalized design and penalty matrices for O'Sullivan spline regression
}
\keyword{internal}
