% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_cusp_reg.R
\name{cusp.prior.mean}
\alias{cusp.prior.mean}
\title{Compute the CUSP prior mean}
\usage{
cusp.prior.mean(t, prior)
}
\arguments{
\item{t}{...}

\item{prior}{...}
}
\value{
A short description...
}
\description{
A short description...
}
\details{
Additional details...
}
