% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_cusp_reg.R
\name{cusp.prior.sim}
\alias{cusp.prior.sim}
\title{Simulate from the prior distribution}
\usage{
cusp.prior.sim(npar, niter, prior)
}
\description{
Simulate from the prior distribution
}
\keyword{internal}
