% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgp_reg_utils.R
\name{get.block.dim}
\alias{get.block.dim}
\title{mgp_reg_utils.R
author: Cristian Castiglione
creation: 04/08/2023
last change: 16/09/2023}
\usage{
get.block.dim(X, Z)
}
\description{
mgp_reg_utils.R
author: Cristian Castiglione
creation: 04/08/2023
last change: 16/09/2023
}
