% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline_basis.R
\name{get.bspline.matrix}
\alias{get.bspline.matrix}
\title{Create the basis matrix for third order B-spline regression}
\usage{
get.bspline.matrix(x, knots, a, b)
}
\description{
Create the basis matrix for third order B-spline regression
}
\keyword{internal}
