% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline_basis.R
\name{get.bspline.penalty}
\alias{get.bspline.penalty}
\title{Create the penalty matrix for third order O'SUllivan spline regression}
\usage{
get.bspline.penalty(knots, a, b)
}
\description{
Create the penalty matrix for third order O'SUllivan spline regression
}
\keyword{internal}
