% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline_basis.R
\name{get.bspline}
\alias{get.bspline}
\title{Create standard design and penalty matrices for B-spline regression}
\usage{
get.bspline(x, n = 20, a = 0, b = 1, unif = FALSE)
}
\arguments{
\item{x}{description}

\item{n}{number of knots to include in the basis expansion}

\item{a}{lower bound of x}

\item{b}{upper bound of x}

\item{unif}{if \code{TRUE} select the knots uniformly in \code{[a,b]} (default \code{FALSE})}
}
\value{
description
\describe{
  \item{\code{k}}{vector of basis knots}
  \item{\code{B}}{B-spline basis matrix}
  \item{\code{P}}{B-spline penalty matrix}
}
}
\description{
\code{get.bspline} compute the basis and penalty matrices for a B-spline basis expansion of dimension \code{n}
}
