% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fourier_basis.R
\name{get.fourier}
\alias{get.fourier}
\title{Create the Fourier design matrix for scatter smoothing}
\usage{
get.fourier(x, n, a = NULL, b = NULL, f = NULL)
}
\arguments{
\item{x}{description}

\item{n}{number of knots to include in the basis expansion}

\item{a}{lower bound of x}

\item{b}{upper bound of x}

\item{f}{frequency value}
}
\description{
\code{get.fourier} compute the basis matrix for a Fourier basis expansion of dimension \code{n}
}
