% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline_basis.R
\name{get.ospline}
\alias{get.ospline}
\title{Create the othogonalized design matrix for O'Sullivan spline regression}
\usage{
get.ospline(x, n = 20, a = 0, b = 1, unif = FALSE, check = TRUE)
}
\arguments{
\item{x}{description}

\item{n}{number of knots to include in the basis expansion}

\item{a}{lower bound of x}

\item{b}{upper bound of x}

\item{unif}{if \code{TRUE}, selects the knots uniformly in \code{[a,b]} (default \code{FALSE})}

\item{check}{if \code{TRUE}, perform a stability check on the orthogonalized basis}
}
\value{
\code{get.ospline} returns a list containig the following elements:
\describe{
  \item{\code{d}}{vector of positive eigenvalues of the B-spline penalty matrix}
  \item{\code{X}}{basis matrix spanning the null-space of the B-spline penalty}
  \item{\code{Z}}{basis matrix spanning the non-linear part of the expansion}
}
}
\description{
\code{get.ospline} compute the basis matrices for an O'Sullivan spline basis expansion,
which consists of an orthogonalized version of B-spline representation where the basis
matrix is rotated and scaled using the eigenvectors and eigenvalues of the B-spline 
penalty matrix.
}
\references{
Wand, Ormerod (2008) 
 On semiparametric regression with O'Sullivan penalized splines
 Australian & New Zealand Journal of Statistics, 50(2): 179-198
}
