% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis_wavelet.R
\name{get.wavelet}
\alias{get.wavelet}
\title{Create a wavelet design matrix of a certain order (default = 6)}
\usage{
get.wavelet(x, rng = range(x), n = 6, f = 5, res = 16384)
}
\arguments{
\item{x}{vector of evaluation points}

\item{rng}{x range}

\item{n}{number of levels}

\item{f}{filter number}

\item{res}{resolution}
}
\value{
\code{W} description
}
\description{
\code{get.wavelet} description
}
\details{
wavelet_test.R
author: Cristian Castiglione
creation: 01/12/2023
last change: 01/12/2023
}
