% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_blm_reg.R
\name{init.blm.prior}
\alias{init.blm.prior}
\title{blm_reg_utils.R
author: Cristian Castiglione
creation: 05/01/2023
last change: 05/01/2023}
\usage{
init.blm.prior(prior)
}
\description{
blm_reg_utils.R
author: Cristian Castiglione
creation: 05/01/2023
last change: 05/01/2023
}
