% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_cusp_reg.R
\name{init.cusp.control}
\alias{init.cusp.control}
\title{Function for initializing the control parameters of the algorithms}
\usage{
init.cusp.control(control)
}
\description{
Function for initializing the control parameters of the algorithms
}
\keyword{internal}
