% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_cusp_reg.R
\name{init.cusp.param}
\alias{init.cusp.param}
\title{Function for initializing the unknown parameters of the model}
\usage{
init.cusp.param(y, C, idx, prior)
}
\description{
Function for initializing the unknown parameters of the model
}
\keyword{internal}
