% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_cusp_reg.R
\name{init.cusp.prior}
\alias{init.cusp.prior}
\title{utils_cusp_reg.R
author: Cristian Castiglione
creation: 05/01/2024
last change: 05/01/2024
Function for initializing the prior parameters of the MGP}
\usage{
init.cusp.prior(prior)
}
\description{
utils_cusp_reg.R
author: Cristian Castiglione
creation: 05/01/2024
last change: 05/01/2024
Function for initializing the prior parameters of the MGP
}
\keyword{internal}
