% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_mgp_reg.R
\name{init.mgp.control}
\alias{init.mgp.control}
\title{Function for initializing the control parameters of the algorithms}
\usage{
init.mgp.control(control)
}
\description{
Function for initializing the control parameters of the algorithms
}
\keyword{internal}
