% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_mgp_reg.R
\name{init.mgp.prior}
\alias{init.mgp.prior}
\title{mgp_reg_utils.R
author: Cristian Castiglione
creation: 04/08/2023
last change: 16/09/2023
Function for initializing the prior parameters of the MGP}
\usage{
init.mgp.prior(prior)
}
\description{
mgp_reg_utils.R
author: Cristian Castiglione
creation: 04/08/2023
last change: 16/09/2023
Function for initializing the prior parameters of the MGP
}
\keyword{internal}
