% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast_rmnorm.R
\name{lschol}
\alias{lschol}
\title{Solve the linear system using the forward-backward Cholesky algorithm}
\usage{
lschol(A, b, upper = FALSE)
}
\arguments{
\item{A}{a square positive definite numeric matrix}

\item{b}{a numeric vector/matrix}

\item{upper}{if \code{TRUE}, uses the upper Cholesky factor of \code{A}}
}
\value{
A list containing:
\describe{
  \item{\code{sol}}{the solution of the linear system}
  \item{\code{chol}}{the upper/lower Cholesky factor of \code{A}}
  \item{\code{inv}}{the inverse Cholesky factor of \code{A}}
}
}
\description{
\code{lschol} implements the \code{LS-CHOL} algorithm to solve the linear system
\eqn{Ax = b} with respect to \eqn{x}, where \eqn{A} is a square positive definite
matrix and \eqn{b} is a vector or a matrix with compatible dimensions.
}
\keyword{internal}
