% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgp_gam_fit.R
\name{mgp.gam.fit}
\alias{mgp.gam.fit}
\title{mgp_gam_fit.R
author: Cristian Castiglione
creation: 15/10/2023
last change: 16/10/2023}
\usage{
mgp.gam.fit(y, X, Z, prior = list(), control = list())
}
\description{
description: 
  This file implements a Gibbs sampling algorithm for estimating a Bayesian 
  additive model predicting the response y given the covariate matrices 
  X, Z_1, ..., Z_H. Assuming that the columns of each Z_h (h = 1, ..., H) are 
  sorted in a decresing order of importance, we elicit for the regreesion 
  coefficients beta_h an increasing shrinkage prior based on the multiplicative 
  gamma process by Bhattacharya & Dunson (2011). See also Durante (2017).
  
references:
  Bhattacharya & Dunson (2011)
  Sparse Bayesian infinite factor models
  Biometrika, 98(2): 291-306
  
  Durante (2017). 
  A note on the multiplicative gamma process
  Statistics and Probability Letters, 122: 198-204
  
model specification:
}
\details{
y = X beta_0 + Z1 beta_1 + ... + Z_H beta_H + e, 
  e ~ N(0, psi^-1), psi ~ Gamma(a, b),
  beta_0 ~ N(0, kappa^-1 I) (kappa -> inf),
  beta_jh ~ N(0, tau_h^-1 phi_jh^-1),
  phi_jh = delta_1h x ... x delta_jh,
  delta_1h ~ Gamma(a1, 1), delta_jh ~ Gamma(a2, 1) (j > 1)
  tau_h ~ Gamma(v/2, v/2),
  
  for j = 1, ..., ph and h = 1, ..., H, where a1, a2, a, b, v > 0 are 
  fixed hyperparameters.
  To induce an increasing shrinkage on beta, we need to impose a2 > a1.
  Moreover, to guarantee for 1 / phi_jh (j = 1, ..., ph) to have finite
  first and second moment, we need a1 > 2 and a3 > 3. 

Take a look to the following articles for an efficient sampling from
the full-conditional distribution of beta:

Bhattacharya, Chakraborty, Mallik (2016)
Fast sampling with Gaussian scale mixture priors in high-dimensional regression
Biometrika, 103(4): 985-991

Rue (2001)
Fast sampling of Gaussian Markov random fields
Journal of the Royal Statistical Society, Series B, 63(): 325-338
}
