% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_mgp_reg.R
\name{mgp.prior.mean}
\alias{mgp.prior.mean}
\title{Compute the MGP prior mean}
\usage{
mgp.prior.mean(t, prior)
}
\arguments{
\item{t}{...}

\item{prior}{...}
}
\value{
A short description...
}
\description{
A short description...
}
\details{
Additional details...
}
