% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast_rmnorm.R
\name{rmnorm}
\alias{rmnorm}
\title{Draw a sample from a multivariate Gaussian distribution}
\usage{
rmnorm(A, b)
}
\arguments{
\item{A}{a square positive definite numeric matrix}

\item{b}{a numeric vector}
}
\value{
A vector containing the sampled Gaussian variables.
}
\description{
\code{rmnorm} draws a sample \eqn{x} from a multivariate Gaussian distribution 
\eqn{N_p(\mu, \Sigma)} having mean vector \eqn{\mu = A^{-1} b} and variance matrix 
\eqn{\Sigma = A^{-1}}, where \eqn{A} is a \eqn{p \times p} positive definite 
matrix and \eqn{b} is a \eqn{p \times 1} vector.
}
\details{
\code{rmnorm} uses a fast sampling scheme based on the algorithm proposed by Rue (2001).
}
\references{
Rue (2001)
  Fast sampling of Gaussian Markov random fields.
  Journal of the Royal Statistical Society, Series B, 63(2): 325-338
}
\keyword{internal}
