% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fourier_basis.R
\name{get.fourier}
\alias{get.fourier}
\title{fourier_basis.R
author: Cristian Castiglione
creation: 10/08/2023
last change: 24/10/2023
reference:
   Ramsay, Silverman (2005) 
   Functional data analysis, Second edition
   Springer}
\usage{
get.fourier(x, n, a = NULL, b = NULL, f = NULL)
}
\description{
fourier_basis.R
author: Cristian Castiglione
creation: 10/08/2023
last change: 24/10/2023
reference:
   Ramsay, Silverman (2005) 
   Functional data analysis, Second edition
   Springer
}
