% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis_fourier.R
\name{get.fourier}
\alias{get.fourier}
\title{Create the Fourier design matrix for scatter smoothing}
\usage{
get.fourier(x, n, a = NULL, b = NULL, f = NULL)
}
\arguments{
\item{x}{description}

\item{n}{number of knots to include in the basis expansion}

\item{a}{lower bound of x}

\item{b}{upper bound of x}

\item{f}{frequency value}
}
\value{
\code{get.fourier} returns a list containing the following elements:
\describe{
  \item{\code{d}}{vector of positive eigenvalues of the Fourier penalty matrix}
  \item{\code{X}}{basis matrix spanning the null-space of the Fourier penalty}
  \item{\code{Z}}{basis matrix spanning the non-linear part of the expansion}
}
}
\description{
\code{get.fourier} compute the basis matrix for a Fourier basis expansion of dimension \code{n}
}
\details{
file: fourier_basis.R
author: Cristian Castiglione
creation: 10/08/2023
last change: 24/10/2023
reference:
   Ramsay, Silverman (2005) 
   Functional data analysis, Second edition
   Springer
}
\references{
Ramsay, Silverman (2005) 
  Functional data analysis, Second edition
  Springer
}
